move exec and exit into their own files
