+++ src/process/clone.h
@@ -1,5 +1,8 @@
 #pragma once
 
+#include "push_message.h"
+#include "helpers.h"
+
 typedef struct {
     u64 pid_tgid;
     clone_info_t clone_info;
