+++ src/script-events.c
@@ -33,15 +33,6 @@ struct bpf_map_def SEC("maps/read_path_skip") read_path_skip = {
     .namespace = "",
 };
 
-struct bpf_map_def SEC("maps/tail_call_table") tail_call_table = {
-    .type = BPF_MAP_TYPE_PROG_ARRAY,
-    .key_size = sizeof(u32),
-    .value_size = sizeof(u32),
-    .max_entries = 32,
-    .pinning = 0,
-    .namespace = "",
-};
-
 struct bpf_map_def SEC("maps/process_ids") process_ids = {
     .type = BPF_MAP_TYPE_HASH,
     .key_size = sizeof(u64),
